+++ cps/about.py
@@ -67,6 +67,7 @@ _VERSIONS = OrderedDict(
     Unidecode = unidecode_version,
     Flask_SimpleLDAP =  u'installed' if bool(services.ldap) else u'not installed',
     Goodreads = u'installed' if bool(services.goodreads_support) else u'not installed',
+    jsonschema = services.SyncToken.__version__  if bool(services.SyncToken) else u'not installed',
 
 )
 _VERSIONS.update(uploader.get_versions())
