+++ cps/templates/generate_kobo_auth_url.html
@@ -1,7 +1,6 @@
-{% extends "layout.html" %}
+{% extends "fragment.html" %}
 {% block body %}
 <div class="well">
-  <h2 style="margin-top: 0">{{_('Generate Kobo Auth URL')}}</h2>
   <p>
     {{_('Open the .kobo/Kobo eReader.conf file in a text editor and add (or edit):')}}</a>.
   </p>
