+++ optional-requirements.txt
@@ -32,3 +32,6 @@ rarfile>=2.7
 # other
 natsort>=2.2.0
 git+https://github.com/OzzieIsaacs/comicapi.git@5346716578b2843f54d522f44d01bc8d25001d24#egg=comicapi
+
+#kobo integration
+jsonschema>=3.2.0
