+++ requirements.txt
@@ -13,4 +13,3 @@ SQLAlchemy>=1.1.0
 tornado>=4.1
 Wand>=0.4.4
 unidecode>=0.04.19
-jsonschema>=3.2.0
